Atmosphere generation: Improved turbulence sharing using file backend. This allows super large screens like twice the physical memory to work without causing swapping during simulation. The screens are generated one by one (two screens in fft method), saved to file. And then mmaped for reading during simulatiosn. The portion of screens that are used will be retrieved from file, while other portion won't be loaded at all.
atm.wdrand now takes a role in seeding wind direction random stream when it is not zero. So we can run the same screen with different wind directions by setting the same seed, but different atm.wdrand number.

map_t is now compatible to dmat so that we can use the methods of dmat to load/save from/to files, or mathematical calculations.

The headers in dmat, etc will be saved to file when they exist.