Overhauled the ray tracing rules. For linear interpolation, the outmost line of grid defines the boundary. For bicubic interpolation, For ray tracing from map_t, and from loc_t and loc_t->map when MAOS_LOC_EXTEND=0 the second outmost line of grid defines the boundary. For ray tracing from loc_t and loc_t->map and when MAOS_LOC_EXTEND=1, the outmost line of grid defines the boundary.
The boundary is defined in such a sense that interpolation results drop to zero outside of the boundary, without any edge
roll off. For a point right on the edge, if it is the left or bottom edge, it receives result from the grid, but for
right or top edges, it receives 0.