1) Revised low order modes for LTAO to include focus and astigmatism. 2) Revised dbg.atm to be more flexible for simulating special turbulence screens.
The reason for 1) is explained here. Imagine if there is a focus mode of strength a (averaged over the pupil) in high altitude turbulence at h. For LGS WFS, it creates the same effect as a*(1-h/hlgs) in ground due to cone effect. The tomography and DM fitting will therefore put this amount of focus to the ground DM. The LGS WFS is happy, but the science and NGS WFS will see a*h/hlgs amount of focus, plus plate scale error. Similar analysis applies to the astigmatism modes. In NFIRAOS, we use low order NGS WFS to sense and fix these in name of plate scale modes. But how do you do it in LTAO? You cannot simply correct such focus/astigmatism modes using NGS because they are not separable from genuine focus/astigmatism modes on the ground. A possible solution would be to use the NGS WFS focus/astigmatism as the "truth measurement" at low frequency by frequency blending NGS and LGS measurements.